refactor(web): extract variant CRUD handlers into useVariantActions
Fourth editor-shell hook. useVariantActions groups the five variant handlers (switch / add / duplicate / delete / rename). Kept deliberately thin: the variants state, the module cache it writes through (setVariants), and the structural epoch stay in the composition root — those carry the subtle remount/race semantics that are clearest co-located. The hook just sequences snapshot to mutate to navigate.

Behaviour-preserving: handler bodies moved verbatim. Verified live in the browser preview — add (URL gains ?v=1, second variant tab, clean remount) and switch (back to variant 0) both work with no error boundary. typecheck + tests (web 89) + lint + fmt green.
